+++ docs/changelog.md
@@ -38,7 +38,8 @@
   ```
   before loading it.
   [#855](https://github.com/iodide-project/pyodide/pull/855)
-
+- Javascript exceptions can now be raised and caught in Python. They are wrapped in pyodide.JsException.
+  [#872](https://github.com/iodide-project/pyodide/pull/872)
 
 ## Version 0.15.0
 *May 19, 2020*